pyproject.toml
[project]
name = "vtodo"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "celery>=5.3",
    "redis>=5.0",
    "django-htmx>=1.17",
    "whitenoise>=6.6",
    "psycopg2-binary>=2.9",
    "gunicorn>=21.0",
    "requests>=2.31",
    "mf2py>=2.0.1",
    "django-environ>=0.11",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "django-debug-toolbar>=4.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
pythonpath = ["."]
